ADD molkky
Introductory exercise to small classes and pointers in C++.
Cmd line based program that can be used as a point counter in mölkky with 2 players. In this game the players are named Matti and Teppo.

Mölkky is a block throwing game where your aim is to get exactly 50 points. If a player gets more than 50 points, their score is reduced to 25 points.

Each turn, the user can input the amount of points scored. After that, the program will print out the scoreboard. Points are automatically reduced to 25 if the player reaches more than 50 points. When either player reaches exactly 50 points, the program will announce the winner.